refactor: 보드 페이지 모듈 및 컴포넌트 별도로 분리
